user add with images fixed with product and category can delete avail

+++ controllers/category.controller.js
@@ -1,23 +1,30 @@
-import { productCate } from "../models/models.js";
+import { productCate, productModel } from "../models/models.js";
 import { AppError, DatabaseError, ServerError } from "../lib/customError.js";
 import { successResponse } from "../utils/apiResponse.js";
-async function createProductCata(req, res, next) {
-  const { categoryName, categoryImage = "" } = req.body;
-  const hadCata = await productCate.find({ categoryName: categoryName });
-  if (hadCata.length > 0) {
-    let userErr = new AppError("Category already in use!!", 400);
-    return next(userErr);
-  }
+import { uploadSingleToCloudinary } from "../middleware/uploadImage.js";
 
+async function createProductCata(req, res, next) {
+  const { name } = req.body;
+  let cateName = JSON.parse(name);
+  const hadCata = await productCate.find({ categoryName: cateName });
+  if (hadCata.length > 0) {
+    let cateErr = new AppError("Category already in use!!", 400);
+    return next(cateErr);
+  }
+  const fileBuffer = req.file.buffer; // File buffer from Multer
+  const folder = "category"; // Cloudinary folder name
+
+  // Upload image to Cloudinary
+  const categoryImageUrl = await uploadSingleToCloudinary(fileBuffer, folder);
   const newCata = new productCate({
-    categoryName,
-    categoryImage,
+    categoryName: cateName,
+    categoryImage: categoryImageUrl?.url,
   });
 
   let savedCata = await newCata.save();
   if (!savedCata) {
-    let userErr = new DatabaseError("Failed to create user!!");
-    return next(userErr);
+    let cateErr = new DatabaseError("Failed to create user!!");
+    return next(cateErr);
   }
 
   return successResponse(res, 201, "Category Created Successfully", savedCata);
@@ -58,6 +65,11 @@ async function updateProductCata(req, res, next) {
 }
 async function deleteProductCata(req, res, next) {
   const { id } = req.params;
+  const productDel = await productModel.deleteMany({ category: id });
+  if (!productDel) {
+    let serverErr = new DatabaseError("failed to delete user!!");
+    return next(serverErr);
+  }
   const deletedCata = await productCate.findByIdAndDelete(id);
   if (!deletedCata) {
     let serverErr = new DatabaseError("failed to delete user!!");
